+++ model/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @platforma-open/milaboratories.mixcr-clonotyping.model
 
+## 1.4.4
+
+### Patch Changes
+
+- 9454329: Update ui-vue to 1.21.21, model to 1.21.20
+
+## 1.4.3
+
+### Patch Changes
+
+- 02b681a: use PlChartStackedBar component
+
 ## 1.4.2
 
 ### Patch Changes