+++ test/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @platforma-open/milaboratories.mixcr-clonotyping.test
 
+## 1.3.4
+
+### Patch Changes
+
+- 9454329: Update ui-vue to 1.21.21, model to 1.21.20
+- Updated dependencies [9454329]
+  - @platforma-open/milaboratories.mixcr-clonotyping.model@1.4.4
+
+## 1.3.3
+
+### Patch Changes
+
+- 02b681a: use PlChartStackedBar component
+- Updated dependencies [02b681a]
+  - @platforma-open/milaboratories.mixcr-clonotyping.model@1.4.3
+
 ## 1.3.2
 
 ### Patch Changes